v1.0.1

+++ crawler/init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `packages` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(11) unsigned NOT NULL AUTO_INCREMENT,
   `name` varchar(128) NOT NULL DEFAULT '',
   `description` text,
   `version` varchar(50) DEFAULT NULL,
@@ -11,9 +11,9 @@ CREATE TABLE IF NOT EXISTS `packages` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE IF NOT EXISTS `downloads` (
-  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
-  `package` int(11) NOT NULL,
-  `downloads` int(11) NOT NULL,
+  `id` bigint(11) unsigned NOT NULL AUTO_INCREMENT,
+  `package` bigint(11) unsigned NOT NULL,
+  `downloads` bigint(11) NOT NULL,
   `day` date NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `DownloadsPerDay` (`package`,`day`),
